refactor(web): remove references to client/storage
