+++ web/src/components/storage/EncryptionSettingsDialog.tsx
@@ -26,7 +26,7 @@ import React, { useEffect, useState, useRef } from "react";
 import { Checkbox, Form, Switch, Stack } from "@patternfly/react-core";
 import { _ } from "~/i18n";
 import { PasswordAndConfirmationInput, Popup } from "~/components/core";
-import { EncryptionMethods } from "~/client/storage";
+import { EncryptionMethods } from "~/types/storage";
 
 const DIALOG_TITLE = _("Encryption");
 const DIALOG_DESCRIPTION = _(
